feat(frontend):add shopping cart

+++ frontend/app/action/cart.ts
@@ -1,13 +1,20 @@
 "use server";
 
 import axios from "axios";
+import { getToken } from "./token";
 
 const GO_API_URL = process.env.GO_API_URL;
 
 export async function getCart() {
   try {
+    const token = await getToken();
+    if (!token) {
+      return { success: false, message: "Not authenticated" };
+    }
     const response = await axios.get(`${GO_API_URL}/getcart`, {
-      withCredentials: true,
+      headers: {
+        Cookie: `jwt=${token}`,
+      },
     });
 
     return {
@@ -24,20 +31,27 @@ export async function getCart() {
 
 export async function addToCart(productId: number, quantity: number = 1) {
   try {
+    const token = await getToken();
+    if (!token) {
+      return { success: false, message: "Not authenticated" };
+    }
+
     const response = await axios.post(
       `${GO_API_URL}/addtocart`,
+      { product_id: productId, quantity },
       {
-        product_id: productId,
-        quantity,
+        headers: {
+          Cookie: `jwt=${token}`,
+        },
       },
-      { withCredentials: true },
     );
 
     return {
       success: true,
       message: response.data.message || "Added to cart",
     };
-  } catch {
+  } catch (error) {
+    console.error("AddToCart error:", error);
     return {
       success: false,
       message: "Failed add to cart",
@@ -47,8 +61,14 @@ export async function addToCart(productId: number, quantity: number = 1) {
 
 export async function deleteCardItem() {
   try {
+    const token = await getToken();
+    if (!token) {
+      return { success: false, message: "Not authenticated" };
+    }
     await axios.delete(`${GO_API_URL}/deletecartitem`, {
-      withCredentials: true,
+      headers: {
+        Cookie: `jwt=${token}`,
+      },
     });
 
     return {
@@ -59,6 +79,21 @@ export async function deleteCardItem() {
     return {
       success: false,
       message: "Failed delete cart",
+    };
+  }
+}
+
+export async function removeCartItem(itemId: number) {
+  try {
+    await axios.delete(`${GO_API_URL}/deletecartitem/${itemId}`);
+    return {
+      success: true,
+      message: "Item deleted",
+    };
+  } catch {
+    return {
+      success: false,
+      message: "Failed delete item",
     };
   }
 }
